feat: refactor admin management forms; rename EditAdminForm component file and update imports; implement EditAdminForm with Zod validation and form handling; enhance AddAdminForm with ChangeEvent typing; remove unused ConfettiButton component
