+++ src/components/pages/admin-management/add-admin-form.tsx
@@ -3,7 +3,7 @@
 import * as z from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
-import { useState } from 'react';
+import { useState, ChangeEvent } from 'react';
 import { Loader2 } from 'lucide-react';
 import { useAdminStore } from '@/hooks/admin/use-admins';
 import { toast } from 'sonner';
@@ -129,7 +129,7 @@ export function AddAdminForm() {
 										<Input
 											placeholder='Input your full name'
 											{...field}
-											onChange={(e) => {
+											onChange={(e: ChangeEvent<HTMLInputElement>) => {
 												field.onChange(e);
 												handleFormChange('name', e.target.value);
 											}}
@@ -150,7 +150,7 @@ export function AddAdminForm() {
 										<Input
 											placeholder='Input your username'
 											{...field}
-											onChange={(e) => {
+											onChange={(e: ChangeEvent<HTMLInputElement>) => {
 												field.onChange(e);
 												handleFormChange('username', e.target.value);
 											}}
@@ -171,7 +171,7 @@ export function AddAdminForm() {
 										<PasswordInput
 											placeholder='Input your password'
 											{...field}
-											onChange={(e) => {
+											onChange={(e: ChangeEvent<HTMLInputElement>) => {
 												field.onChange(e);
 												handleFormChange('password', e.target.value);
 											}}
@@ -198,7 +198,7 @@ export function AddAdminForm() {
 												className='z-10 shadow-none -ms-px rounded-s-none'
 												placeholder='Input your phone number'
 												{...field}
-												onChange={(e) => {
+												onChange={(e: ChangeEvent<HTMLInputElement>) => {
 													field.onChange(e);
 													handleFormChange('phone', e.target.value);
 												}}
@@ -222,7 +222,7 @@ export function AddAdminForm() {
 											type='email'
 											placeholder='Input your email'
 											{...field}
-											onChange={(e) => {
+											onChange={(e: ChangeEvent<HTMLInputElement>) => {
 												field.onChange(e);
 												handleFormChange('email', e.target.value);
 											}}
