+++ src/components/ui/confetti.tsx
@@ -15,7 +15,6 @@ import type {
 } from 'canvas-confetti';
 import confetti from 'canvas-confetti';
 
-import { Button, ButtonProps } from '@/components/ui/button';
 
 type Api = {
 	fire: (options?: ConfettiOptions) => void;
@@ -93,33 +92,6 @@ const Confetti = forwardRef<ConfettiRef, Props>((props, ref) => {
 	);
 });
 
-interface ConfettiButtonProps extends ButtonProps {
-	options?: ConfettiOptions &
-		ConfettiGlobalOptions & { canvas?: HTMLCanvasElement };
-	children?: React.ReactNode;
-}
-
-function ConfettiButton({ options, children, ...props }: ConfettiButtonProps) {
-	const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-		const rect = event.currentTarget.getBoundingClientRect();
-		const x = rect.left + rect.width / 2;
-		const y = rect.top + rect.height / 2;
-		confetti({
-			...options,
-			origin: {
-				x: x / window.innerWidth,
-				y: y / window.innerHeight,
-			},
-		});
-	};
-
-	return (
-		<Button onClick={handleClick} {...props}>
-			{children}
-		</Button>
-	);
-}
-
 Confetti.displayName = 'Confetti';
 
-export { Confetti, ConfettiButton };
+export { Confetti };
