+++ src/components/ui/password-input.tsx
@@ -4,60 +4,56 @@ import * as React from 'react';
 import { EyeIcon, EyeOffIcon } from 'lucide-react';
 
 import { Button } from '@/components/ui/button';
-import { Input, InputProps } from '@/components/ui/input';
+import { Input } from '@/components/ui/input';
 import { cn } from '@/lib/utils';
 
-const PasswordInput = (
-    {
-        ref,
-        className,
-        ...props
-    }: InputProps & {
-        ref: React.RefObject<HTMLInputElement>;
-    }
-) => {
-    const [showPassword, setShowPassword] = React.useState(false);
-    const disabled =
-        props.value === '' || props.value === undefined || props.disabled;
+export type PasswordInputProps = React.InputHTMLAttributes<HTMLInputElement>;
 
-    return (
-        <div className='relative'>
-            <Input
-                type={showPassword ? 'text' : 'password'}
-                className={cn('hide-password-toggle pr-10', className)}
-                ref={ref}
-                {...props}
-            />
-            <Button
-                type='button'
-                variant='ghost'
-                size='sm'
-                className='top-0 right-0 absolute hover:bg-transparent px-3 py-2 h-full'
-                onClick={() => setShowPassword((prev) => !prev)}
-                disabled={disabled}
-            >
-                {showPassword && !disabled ? (
-                    <EyeIcon className='w-4 h-4' aria-hidden='true' />
-                ) : (
-                    <EyeOffIcon className='w-4 h-4' aria-hidden='true' />
-                )}
-                <span className='sr-only'>
-                    {showPassword ? 'Hide password' : 'Show password'}
-                </span>
-            </Button>
+const PasswordInput = React.forwardRef<HTMLInputElement, PasswordInputProps>(
+	({ className, ...props }, ref) => {
+		const [showPassword, setShowPassword] = React.useState(false);
+		const disabled =
+			props.value === '' || props.value === undefined || props.disabled;
 
-            {/* hides browsers password toggles */}
-            <style>{`
-                .hide-password-toggle::-ms-reveal,
-                .hide-password-toggle::-ms-clear {
-                    visibility: hidden;
-                    pointer-events: none;
-                    display: none;
-                }
-            `}</style>
-        </div>
-    );
-};
+		return (
+			<div className='relative'>
+				<Input
+					type={showPassword ? 'text' : 'password'}
+					className={cn('hide-password-toggle pr-10', className)}
+					ref={ref}
+					{...props}
+				/>
+				<Button
+					type='button'
+					variant='ghost'
+					size='sm'
+					className='top-0 right-0 absolute hover:bg-transparent px-3 py-2 h-full'
+					onClick={() => setShowPassword((prev) => !prev)}
+					disabled={disabled}
+				>
+					{showPassword && !disabled ? (
+						<EyeIcon className='w-4 h-4' aria-hidden='true' />
+					) : (
+						<EyeOffIcon className='w-4 h-4' aria-hidden='true' />
+					)}
+					<span className='sr-only'>
+						{showPassword ? 'Hide password' : 'Show password'}
+					</span>
+				</Button>
+
+				{/* hides browsers password toggles */}
+				<style>{`
+                    .hide-password-toggle::-ms-reveal,
+                    .hide-password-toggle::-ms-clear {
+                        visibility: hidden;
+                        pointer-events: none;
+                        display: none;
+                    }
+                `}</style>
+			</div>
+		);
+	},
+);
 PasswordInput.displayName = 'PasswordInput';
 
 export { PasswordInput };
